fixup! [LibOS] Simplify `getdents/getdents64`

+++ LibOS/shim/include/shim_table.h
@@ -100,7 +100,7 @@ long shim_do_fsync(int fd);
 long shim_do_fdatasync(int fd);
 long shim_do_truncate(const char* path, loff_t length);
 long shim_do_ftruncate(int fd, loff_t length);
-long shim_do_getdents(int fd, struct linux_dirent* buf, size_t count);
+long shim_do_getdents(int fd, struct linux_dirent* buf, unsigned int count);
 long shim_do_getcwd(char* buf, size_t size);
 long shim_do_chdir(const char* filename);
 long shim_do_fchdir(int fd);
